chore(utils): move encoding utils and process to their respective modules
